fix: change price fields from double to BigDecimal for improved precision

+++ src/main/java/net/andrecarbajal/sysped/model/Order.java
@@ -4,6 +4,7 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -33,7 +34,7 @@ public class Order {
     private OrderStatus status = OrderStatus.PENDIENTE;
 
     @Column(nullable = false)
-    private double priceTotal;
+    private BigDecimal priceTotal;
 
     @OneToMany(mappedBy = "order", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<OrderDetails> details = new ArrayList<>();
